// src/app/types/AdminInterfaces.ts
export interface AdminInfoDashboard {  
    num_ventas: number;
    monto_ventas: number;
    num_clientes: number;
    monto_ingresos: number;
    num_ingresos: number;
    num_productos: number;
    num_categorias: number;
    num_usuarios: number;

}


/*
{
    "num_ventas": 10,
    "monto_ventas": 100,
    "num_clientes": 5,
    "monto_ingresos": 1000,
    "num_ingresos": 10,
    "num_productos": 10,
    "num_categorias": 5,
    "num_usuarios": 5
}
*/